python-fastapi-stock-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
grpcio
grpcio-tools